requirements.txt
# Pinecone Knowledge Base Requirements
pinecone-client==3.0.3
openai==1.35.0
google-api-python-client==2.112.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pandas==2.1.4
python-dotenv==1.0.0
